+++ src/envs/box2d/__init__.py
@@ -1,13 +1,13 @@
-from src.envs.box2d.meta_lunarlander import CARLLunarLanderEnv
-from src.envs.box2d.meta_vehicle_racing import CARLVehicleRacingEnv
-from src.envs.box2d.meta_bipedal_walker import CARLBipedalWalkerEnv
+from src.envs.box2d.carl_lunarlander import CARLLunarLanderEnv
+from src.envs.box2d.carl_vehicle_racing import CARLVehicleRacingEnv
+from src.envs.box2d.carl_bipedal_walker import CARLBipedalWalkerEnv
 
 # Contextenvs.s and bounds by name
-from src.envs.box2d.meta_lunarlander import DEFAULT_CONTEXT as CARLLunarLanderEnv_defaults
-from src.envs.box2d.meta_lunarlander import CONTEXT_BOUNDS as CARLLunarLanderEnv_bounds
+from src.envs.box2d.carl_lunarlander import DEFAULT_CONTEXT as CARLLunarLanderEnv_defaults
+from src.envs.box2d.carl_lunarlander import CONTEXT_BOUNDS as CARLLunarLanderEnv_bounds
 
-from src.envs.box2d.meta_vehicle_racing import DEFAULT_CONTEXT as CARLVehicleRacingEnv_defaults
-from src.envs.box2d.meta_vehicle_racing import CONTEXT_BOUNDS as CARLVehicleRacingEnv_bounds
+from src.envs.box2d.carl_vehicle_racing import DEFAULT_CONTEXT as CARLVehicleRacingEnv_defaults
+from src.envs.box2d.carl_vehicle_racing import CONTEXT_BOUNDS as CARLVehicleRacingEnv_bounds
 
-from src.envs.box2d.meta_bipedal_walker import DEFAULT_CONTEXT as CARLBipedalWalkerEnv_defaults
-from src.envs.box2d.meta_bipedal_walker import CONTEXT_BOUNDS as CARLBipedalWalkerEnv_bounds
+from src.envs.box2d.carl_bipedal_walker import DEFAULT_CONTEXT as CARLBipedalWalkerEnv_defaults
+from src.envs.box2d.carl_bipedal_walker import CONTEXT_BOUNDS as CARLBipedalWalkerEnv_bounds
