+++ src/envs/brax/__init__.py
@@ -1,25 +1,25 @@
-from src.envs.brax.meta_ant import CARLAnt
-from src.envs.brax.meta_halfcheetah import CARLHalfcheetah
-from src.envs.brax.meta_humanoid import CARLHumanoid
-from src.envs.brax.meta_fetch import CARLFetch
-from src.envs.brax.meta_grasp import CARLGrasp
-from src.envs.brax.meta_ur5e import CARLUr5e
+from src.envs.brax.carl_ant import CARLAnt
+from src.envs.brax.carl_halfcheetah import CARLHalfcheetah
+from src.envs.brax.carl_humanoid import CARLHumanoid
+from src.envs.brax.carl_fetch import CARLFetch
+from src.envs.brax.carl_grasp import CARLGrasp
+from src.envs.brax.carl_ur5e import CARLUr5e
 
 # Contexts and bounds by name
-from src.envs.brax.meta_ant import DEFAULT_CONTEXT as CARLAnt_defaults
-from src.envs.brax.meta_ant import CONTEXT_BOUNDS as CARLAnt_bounds
+from src.envs.brax.carl_ant import DEFAULT_CONTEXT as CARLAnt_defaults
+from src.envs.brax.carl_ant import CONTEXT_BOUNDS as CARLAnt_bounds
 
-from src.envs.brax.meta_halfcheetah import DEFAULT_CONTEXT as CARLHalfcheetah_defaults
-from src.envs.brax.meta_halfcheetah import CONTEXT_BOUNDS as CARLHalfcheetah_bounds
+from src.envs.brax.carl_halfcheetah import DEFAULT_CONTEXT as CARLHalfcheetah_defaults
+from src.envs.brax.carl_halfcheetah import CONTEXT_BOUNDS as CARLHalfcheetah_bounds
 
-from src.envs.brax.meta_humanoid import DEFAULT_CONTEXT as CARLHumanoid_defaults
-from src.envs.brax.meta_humanoid import CONTEXT_BOUNDS as CARLHumanoid_bounds
+from src.envs.brax.carl_humanoid import DEFAULT_CONTEXT as CARLHumanoid_defaults
+from src.envs.brax.carl_humanoid import CONTEXT_BOUNDS as CARLHumanoid_bounds
 
-from src.envs.brax.meta_fetch import DEFAULT_CONTEXT as CARLFetch_defaults
-from src.envs.brax.meta_fetch import CONTEXT_BOUNDS as CARLFetch_bounds
+from src.envs.brax.carl_fetch import DEFAULT_CONTEXT as CARLFetch_defaults
+from src.envs.brax.carl_fetch import CONTEXT_BOUNDS as CARLFetch_bounds
 
-from src.envs.brax.meta_grasp import DEFAULT_CONTEXT as CARLGrasp_defaults
-from src.envs.brax.meta_grasp import CONTEXT_BOUNDS as CARLGrasp_bounds
+from src.envs.brax.carl_grasp import DEFAULT_CONTEXT as CARLGrasp_defaults
+from src.envs.brax.carl_grasp import CONTEXT_BOUNDS as CARLGrasp_bounds
 
-from src.envs.brax.meta_ur5e import DEFAULT_CONTEXT as CARLUr5e_defaults
-from src.envs.brax.meta_ur5e import CONTEXT_BOUNDS as CARLUr5e_bounds
+from src.envs.brax.carl_ur5e import DEFAULT_CONTEXT as CARLUr5e_defaults
+from src.envs.brax.carl_ur5e import CONTEXT_BOUNDS as CARLUr5e_bounds
