+++ src/envs/classic_control/__init__.py
@@ -1,21 +1,21 @@
-from src.envs.classic_control.meta_mountaincar import CARLMountainCarEnv
-from src.envs.classic_control.meta_pendulum import CARLPendulumEnv
-from src.envs.classic_control.meta_acrobot import CARLAcrobotEnv
-from src.envs.classic_control.meta_cartpole import CARLCartPoleEnv
-from src.envs.classic_control.meta_mountaincarcontinuous import CARLMountainCarContinuousEnv
+from src.envs.classic_control.carl_mountaincar import CARLMountainCarEnv
+from src.envs.classic_control.carl_pendulum import CARLPendulumEnv
+from src.envs.classic_control.carl_acrobot import CARLAcrobotEnv
+from src.envs.classic_control.carl_cartpole import CARLCartPoleEnv
+from src.envs.classic_control.carl_mountaincarcontinuous import CARLMountainCarContinuousEnv
 
 # Contexts and bounds by name
-from src.envs.classic_control.meta_acrobot import DEFAULT_CONTEXT as CARLAcrobotEnv_defaults
-from src.envs.classic_control.meta_acrobot import CONTEXT_BOUNDS as CARLAcrobotEnv_bounds
+from src.envs.classic_control.carl_acrobot import DEFAULT_CONTEXT as CARLAcrobotEnv_defaults
+from src.envs.classic_control.carl_acrobot import CONTEXT_BOUNDS as CARLAcrobotEnv_bounds
 
-from src.envs.classic_control.meta_cartpole import DEFAULT_CONTEXT as CARLCartPoleEnv_defaults
-from src.envs.classic_control.meta_cartpole import CONTEXT_BOUNDS as CARLCartPoleEnv_bounds
+from src.envs.classic_control.carl_cartpole import DEFAULT_CONTEXT as CARLCartPoleEnv_defaults
+from src.envs.classic_control.carl_cartpole import CONTEXT_BOUNDS as CARLCartPoleEnv_bounds
 
-from src.envs.classic_control.meta_pendulum import DEFAULT_CONTEXT as CARLPendulumEnv_defaults
-from src.envs.classic_control.meta_pendulum import CONTEXT_BOUNDS as CARLPendulumEnv_bounds
+from src.envs.classic_control.carl_pendulum import DEFAULT_CONTEXT as CARLPendulumEnv_defaults
+from src.envs.classic_control.carl_pendulum import CONTEXT_BOUNDS as CARLPendulumEnv_bounds
 
-from src.envs.classic_control.meta_mountaincar import DEFAULT_CONTEXT as CARLMountainCarEnv_defaults
-from src.envs.classic_control.meta_mountaincar import CONTEXT_BOUNDS as CARLMountainCarEnv_bounds
+from src.envs.classic_control.carl_mountaincar import DEFAULT_CONTEXT as CARLMountainCarEnv_defaults
+from src.envs.classic_control.carl_mountaincar import CONTEXT_BOUNDS as CARLMountainCarEnv_bounds
 
-from src.envs.classic_control.meta_mountaincarcontinuous import DEFAULT_CONTEXT as CARLMountainCarContinuousEnv_defaults
-from src.envs.classic_control.meta_mountaincarcontinuous import CONTEXT_BOUNDS as CARLMountainCarContinuousEnv_bounds
+from src.envs.classic_control.carl_mountaincarcontinuous import DEFAULT_CONTEXT as CARLMountainCarContinuousEnv_defaults
+from src.envs.classic_control.carl_mountaincarcontinuous import CONTEXT_BOUNDS as CARLMountainCarContinuousEnv_bounds
