+++ src/envs/mario/__init__.py
@@ -1,4 +1,4 @@
-from src.envs.mario.meta_mario_env import CARLMarioEnv
+from src.envs.mario.carl_mario_env import CARLMarioEnv
 
-from src.envs.mario.meta_mario_env import DEFAULT_CONTEXT as CARLMarioEnv_defaults
-from src.envs.mario.meta_mario_env import CONTEXT_BOUNDS as CARLMarioEnv_bounds
+from src.envs.mario.carl_mario_env import DEFAULT_CONTEXT as CARLMarioEnv_defaults
+from src.envs.mario.carl_mario_env import CONTEXT_BOUNDS as CARLMarioEnv_bounds
