+++ src/envs/rna/__init__.py
@@ -1,3 +1,3 @@
-from src.envs.rna.rna_env import CARLRnaDesignEnv
-from src.envs.rna.rna_env import DEFAULT_CONTEXT as CARLRnaDesignEnv_defaults
-from src.envs.rna.rna_env import CONTEXT_BOUNDS as CARLRnaDesignEnv_bounds
+from src.envs.rna.carl_rna_env import CARLRnaDesignEnv
+from src.envs.rna.carl_rna_env import DEFAULT_CONTEXT as CARLRnaDesignEnv_defaults
+from src.envs.rna.carl_rna_env import CONTEXT_BOUNDS as CARLRnaDesignEnv_bounds
